Visualization Capabilities

There are various visualization options that are shipped with the package. For instance, we can plot a geometric constraint system by calling

Main.DeformationPaths.plot — Function

plot(F[, filename])

Plot the geometric constraint system F.

This is a wrapper method for plotting of the individual geometric constraint systems. It calls one of the following: plot_flexes!, plot_framework!, plot_frameworkonsurface!, plot_hypergraph!, plot_polytope!, plot_spherepacking! and plot_sphericaldiskpacking!.

Main.DeformationPaths.plot! — Method

plot!(ax, F)

Add a plot on top of the existing axis ax.

A specific continuous motion of a geometric constraint system can be animated via

Main.DeformationPaths.animate — Function

animate(D, F[, filename])

Create an animation of a geometric constraint system F given a previously computed deformation path.

This is a wrapper method for animations for the individual geometric constraint systems. It calls one of the following: animate2D_framework, animate3D_framework, animate3D_frameworkonsurface, animate2D_hypergraph, animate3D_polytope, animate2D_diskpacking, animate3D_spherepacking and animate3D_sphericaldiskpacking.

And we can also visualize a random projection to 2D or 3D of deformation paths:

Main.DeformationPaths.project_deformation_random — Function

project_deformation_random(D, F, projected_dimension[, filename])

Compute a random projection of deformation paths.

This method can either take a single deformation path or a vector of deformation paths and projects it to curves in 2D or 3D. This makes it possible to visualize high-dimensional deformation spaces.

Plotting Capabilities

Main.DeformationPaths.add_shadow! — Function

add_shadow!(ax, F, D[, flex_color, flex_scale, linewidth, arrowsize])

Plot a curve that traces the motion_matrices along the deformation path D.

Main.DeformationPaths.plot_flexes! — Function

plot_flexes!(ax, F, flex_Real, flex_color, flex_scale, linewidth, arrowsize)

Add infinitesimal flexes to a plot of a geometric constraint system

Main.DeformationPaths.plot_framework! — Function

plot_framework!(ax, F)

Plot a bar-joint or angular framework.

Main.DeformationPaths.plot_frameworkonsurface! — Function

plot_frameworkonsurface!(ax, F)

Plot a bar-joint framework constrained to a surface.

Main.DeformationPaths.plot_spherepacking! — Function

plot_spherepacking!(ax, F)

Plot a sphere packing.

Main.DeformationPaths.plot_sphericaldiskpacking! — Function

plot_sphericaldiskpacking!(ax, F)

Plot a spherical disk packing.

Main.DeformationPaths.plot_hypergraph! — Function

plot_hypergraph!(ax, F)

Plot a volume-constrained hypergraph.

Main.DeformationPaths.plot_polytope! — Function

plot_polytope!(ax, F)

Plot a polytope.

Animation Methods

Main.DeformationPaths.animate — Function

animate(F[, filename; flex_mult, random_flex, num_steps, step_size])

Create an animation of a geometric constraint system F.

Main.DeformationPaths.animate2D_framework — Method

animate2D_framework(D, F, filename)

Compute an animation for a 2-dimensional bar-joint framework.

Main.DeformationPaths.animate3D_framework — Method

animate3D_framework(D, F, filename)

Compute an animation for a 3-dimensional bar-joint framework.

Main.DeformationPaths.animate3D_frameworkonsurface — Method

animate3D_frameworkonsurface(D, F, filename)

Compute an animation for a 3-dimensional bar-joint framework constrained to a surface.

Main.DeformationPaths.animate2D_hypergraph — Method

animate2D_hypergraph(D, F, filename)

Compute an animation for a 2-dimensional volume hypergraph.

Main.DeformationPaths.animate3D_polytope — Function

animate3D_polytope(D, F, filename)

Compute an animation for a 3-dimensional polytope.

Main.DeformationPaths.animate2D_diskpacking — Method

animate2D_diskpacking(D, F, filename)

Compute an animation for a 2-dimensional sticky disk packing.

Main.DeformationPaths.animate3D_spherepacking — Method

animate3D_spherepacking(D, F, filename)

Compute an animation for a 3-dimensional sticky sphere packing.

Main.DeformationPaths.animate3D_sphericaldiskpacking — Method

animate3D_sphericaldiskpacking(D, F, filename)

Compute an animation for a disk packing on the 2-sphere in the Minkowski metric.