Auxiliary Functions

Rigidity Checks

Main.DeformationPaths.is_rigid — Method

is_rigid(F[; tol, tol, tested_random_flexes, symmetric_newton])

Heuristically checks if a geometric constraint system F is (continuously) rigid.

Main.DeformationPaths.is_inf_rigid — Method

is_inf_rigid(F[; tol_rank_drop])

Checks if a geometric constraint system F is infinitesimally rigid.

Main.DeformationPaths.is_second_order_rigid — Method

is_second_order_rigid(F)

Checks if a geometric constraint system F is second-order rigid.

See also compute_nonblocked_flex

Predictor-Corrector Methods

Main.DeformationPaths.euler_step — Method

euler_step(G, step_size, prev_flex, point, K_n)

Euler step predicting the next point along the approximate motion.

Returns

predicted_point::Vector{<:Real}: The next point predicted by Euler's method.

predicted_inf_flex::Vector{<:Real}: The tangent vector predicted by Euler's method.

Main.DeformationPaths.newton_correct — Method

newton_correct(G, point[; tol, time_penalty])

Apply Newton's method to correct point back to the constraints in G.

Arguments

G::ConstraintSystem: The underlying geometric constraint system.

point::Vector{<:Real}: The initial point that Newton's method is applied to.

tol::Real (optional): Numerical tolerance that is used as a stopping criterion for Newton's method. Default value: 1e-13.

time_penalty::Real (optional): If Newton's method takes too long, we stop the iteration and throw an error. Here, "too long" is measured in terms of length(point)/time_penalty seconds. Default value: 2.

Returns

q::Vector{<:Real}: A point q such that the Euclidean norm of the evaluated equations is at most tol

See also newton_correct

Main.DeformationPaths.newton_correct — Method

newton_correct(equations, variables, jac, point[; tol, time_penalty])

Apply Newton's method to correct point back to the constraints in equations.

Arguments

equations::Vector{Expression}: Equations to correct point to.

variables::Vector{Variable}: Variables from the affine coordinate ring.

jac::Matrix{Expression}: Jacobian matrix corresponding to equations and variables.

point::Vector{<:Real}: The initial point that Newton's method is applied to.

tol::Real (optional): Numerical tolerance that is used as a stopping criterion for Newton's method. Default value: 1e-13.

time_penalty::Real (optional): If Newton's method takes too long, we stop the iteration and throw an error. Here, "too long" is measured in terms of length(point)/time_penalty seconds. Default value: 2.

Returns

q::Vector{<:Real}: A point q such that the Euclidean norm of the evaluated equations is at most tol

Main.DeformationPaths.symmetric_newton_correct — Method

symmetric_newton_correct(G, point[; tol, time_penalty])

Apply symmetric Newton's method to correct point back to the constraints in equations.

The symmetric Newton corrector evaluates the Jacobian matrix less often.

Arguments

G::ConstraintSystem: The underlying geometric constraint system.

point::Vector{<:Real}: The initial point that Newton's method is applied to.

tol::Real (optional): Numerical tolerance that is used as a stopping criterion for Newton's method. Default value: 1e-13.

time_penalty::Real (optional): If Newton's method takes too long, we stop the iteration and throw an error. Here, "too long" is measured in terms of length(point)/time_penalty seconds. Default value: 2.

Returns

q::Vector{<:Real}: A point q such that the Euclidean norm of the evaluated equations is at most tol

See also symmetric_newton_correct

Main.DeformationPaths.symmetric_newton_correct — Method

symmetric_newton_correct(equations, variables, jac, point[; tol, time_penalty])

Apply symmetric Newton's method to correct point back to the constraints in equations.

The symmetric Newton corrector evaluates the Jacobian matrix less often.

Arguments

equations::Vector{Expression}: Equations to correct point to.

variables::Vector{Variable}: Variables from the affine coordinate ring.

jac::Matrix{Expression}: Jacobian matrix corresponding to equations and variables.

point::Vector{<:Real}: The initial point that Newton's method is applied to.

tol::Real (optional): Numerical tolerance that is used as a stopping criterion for Newton's method. Default value: 1e-13.

time_penalty::Real (optional): If Newton's method takes too long, we stop the iteration and throw an error. Here, "too long" is measured in terms of length(point)/time_penalty seconds. Default value: 2.

Returns

q::Vector{<:Real}: A point q such that the Euclidean norm of the evaluated equations is at most tol

Infinitessimal Flexes

Main.DeformationPaths.compute_nontrivial_inf_flexes — Method

compute_nontrivial_inf_flexes(G, point, K_n[; tol])

Compute the nontrivial infinitesimal flexes of a geometric constraint system G in point.

Main.DeformationPaths.compute_nonblocked_flex — Method

compute_nonblocked_flex(F[; tol_rank_drop, tol])

Compute an infinitesimal flex of F that is not blocked by an equilibrium stress.

Transformation Methods

Main.DeformationPaths.to_Array — Method

to_Array(G, p)

Transform a realization p to a vector of coordinates.

Main.DeformationPaths.to_Array — Method

to_Array(F, p)

Transform a realization p to a vector of coordinates.

Main.DeformationPaths.to_Array — Method

to_Array(F, p)

Transform a realization p to a vector of coordinates.

Main.DeformationPaths.to_Matrix — Method

to_Matrix(G, q)

Transform a vector of coordinates q to a realization matrix.

Main.DeformationPaths.to_Matrix — Method

to_Matrix(F, q)

Transform a vector of coordinates q to a realization matrix.

Special Polytope Methods

Main.DeformationPaths.fix_antipodals! — Method

fix_antipodals!(F)

Entangles the antipodal points in a polytope F so that their position is constrained to antipodal points on a sphere.

Main.DeformationPaths.tetrahedral_symmetry! — Method

tetrahedral_symmetry!(F)

Add constraints to a polytope F representing a tetrahedral symmetry.

Main.DeformationPaths.triangle_shrinking — Method

triangle_shrinking(F)

Evenly shrink the triangular facets of a given polytope and compute the nontrivial infinitesimal flexes in each step.