Constraint System Constructors

Main.DeformationPaths.ConstraintSystem — Type

Class for Constructing a general constraint system.

Attributes

vertices::Vector{Int}: Vertices describing the geometric constraint system. They are given as a list of integers.

variables::Vector{Variable}: Coordinate variables representing the positions of the vertices.

equations::Vector{Expression}: Equations corresponding to the constraint system.

realization::Matrix{<:Real}: Realization of the geometric constraint system satisfying the equations.

jacobian::Matrix{Expression}: Jacobian matrix for the equations and variables.

dimension::Int: Dimension in which the geometric constraint system lives.

xs::Union{Matrix{Variable}, Matrix{Expression}}: Matrix representing the possible realizations of a geometric constraint system.

system::System: Polynomial system consisting of the equations and variables in the form of HomotopyContinuation.

pinned_vertices::Vector{Int}: Pinned vertices of the system. These vertices remain unchanged and are added as constraints. Pinning can, for instance, be used to factor out rigid motions.

Main.DeformationPaths.ConstraintSystem — Method

ConstraintSystem(vertices::Vector{Int}, variables::Vector{Variable}, equations::Vector{Expression}, realization::Matrix{<:Real}, xs)

Constructor of a ConstraintSystem object.

Types of Geometric Constraint Systems

Main.DeformationPaths.SpherePacking — Type

SpherePacking([vertices,] radii, realization[; pinned_vertices])

Class for sticky sphere packings.

Main.DeformationPaths.SphericalDiskPacking — Type

SphericalDiskPacking([vertices,] contacts, [inversive_distances,] realization[; pinned_vertices])

Class for spherical disk packings in Minkowski space.

Main.DeformationPaths.Polytope — Type

Polytope([vertices,] facets, realization[; pinned_vertices])

Class for 3-dimensional polytopes with edge-length and facet planarity constraints.

For the computation of the normals, we assume that the origin lies in the barycenter. Otherwise, we translate the polytope accordingly.

Main.DeformationPaths.FacetPolytope — Type

FacetPolytope([vertices,] facets, realization[; pinned_vertices])

Class for 3-dimensional polytopes with facet planarity constraints.

For the computation of the normals, we assume that the origin lies in the barycenter. Otherwise, we translate the polytope accordingly.

Main.DeformationPaths.FrameworkOnSurface — Type

FrameworkOnSurface([vertices,] bars, realization, surface[; pinned_vertices])

Class for bar-joint frameworks constrained to a surface.

Main.DeformationPaths.Framework — Type

Framework([vertices,] bars, realization[; pinned_vertices])

Class for bar-joint frameworks.

Main.DeformationPaths.VolumeHypergraph — Type

VolumeHypergraph([vertices,] volumes, realization[; pinned_vertices])

Class for volume-constrained hypergraphs.

Main.DeformationPaths.BodyHinge — Type

BodyHinge([vertices,] facets, realization[; pinned_vertices])

Class for body-hinge frameworks, which are essentially polytopes with rigid facets.

Main.DeformationPaths.BodyBar — Type

BodyBar([vertices,] facets, realization[; pinned_vertices])

Class for body-bar frameworks, which are essentially polytopes with rigid facets and connecting .

Main.DeformationPaths.AngularFramework — Type

AngularFramework([vertices,] angles, realization[; pinned_vertices])

Class for angular bar-joint frameworks.

Transformations

Main.DeformationPaths.equations! — Method

equations!(G, equations)

Set the equations of G to equations.

Main.DeformationPaths.equations! — Method

equations!(F, equations)

Set the equations of the geometric constraint system F.

Main.DeformationPaths.add_equations! — Method

add_equations!(G, equations)

Add the equations to those of G.

Main.DeformationPaths.add_equations! — Method

add_equations!(F, equations)

Add equations to the geometric constraint system F.

Main.DeformationPaths.realization! — Method

realization!(G, realization)

Set the realization of the constraint system G to realization.

Main.DeformationPaths.realization! — Method

realization!(F, realization)

Set the realization of the geometric constraint system F to realization.

Main.DeformationPaths.realization! — Method

realization!(F, realization)

Set the realization of the geometric constraint system F to realization.